flytt DTOer inn i klient
